Add endpoint for get transaction with the daemon rpc
Creates a new client type `DaemonRpcClient` to interface with the
daemon's raw rpc endpoints. The client is slightly refactored to cater
to this new client that does not use JsonRpc.

The `get_transactions` method returns a `TransactionResponse` struct.
This struct can then be further deserialized to access the hex or json
encoded Transaction data.